backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
chromadb==1.5.0
openai==2.21.0
python-dotenv==1.2.1
markdown==3.8
slowapi==0.1.9
